MCA6-5 : code prototype for post and get request
