+++ assets/css/popuo-box.css
@@ -996,7 +996,7 @@ button::-moz-focus-inner {
 }
 
 .video-information-text p {
-    color: ##525254;
+    color: #525254;
     margin: 1em 0;
     font-size: 14px;
     line-height: 1.8em;
