+++ package.json
@@ -13,6 +13,7 @@
         "ejs": "^2.6.1",
         "express": "^4.16.4",
         "mysql": "^2.16.0",
+        "path": "^0.12.7",
         "supervisor": "^0.12.0"
     },
     "devDependencies": {}
