+++ views/pages/index.ejs
@@ -24,11 +24,6 @@
                         <div class="recommended-info">
                             <div class="heading">
                                 <h3>MCA Videos</h3>
-                                <ul>
-                                    <li>
-                                        <%= myData.name %>
-                                    </li>
-                                </ul>
 
                             </div>
