+++ views/partials/nav.ejs
@@ -22,10 +22,11 @@
             <div id="small-dialog2" class="mfp-hide">
                 <h3>Create Account</h3>
                 <div class="signup">
-                    <form>
-                        <input type="text" class="email" placeholder="Email" required="required" pattern="([\w-\.]+@([\w-]+\.)+[\w-]{2,4})?" title="Enter a valid email" />
-                        <input type="password" placeholder="Password" required="required" pattern=".{6,}" title="Minimum 6 characters required" autocomplete="off" />
-                        <input type="text" class="email" placeholder="Mobile Number" maxlength="10" pattern="[1-9]{1}\d{9}" title="Enter a valid mobile number" />
+                    <form action="/sign/regProcess" method="POST">
+                        <input type="text" name="name" placeholder="Name" value="saideep" required="required" title="Enter your full name" />
+                        <input type="text" name="email" placeholder="Email" value="saideep.kankarla@gmail.com" required="required" pattern="([\w-\.]+@([\w-]+\.)+[\w-]{2,4})?" title="Enter a valid email" />
+                        <input type="password" name="password" placeholder="Password" value="12345678" required="required" pattern=".{6,}" title="Minimum 6 characters required" autocomplete="off" />
+                        <input type="text" name="phone" placeholder="Mobile Number" value="1234567890" maxlength="10" pattern="[1-9]{1}\d{9}" title="Enter a valid mobile number" />
                         <input type="submit" value="Sign Up" />
                     </form>
                 </div>
@@ -33,17 +34,6 @@
                 <div class="clearfix"> </div>
             </div>
 
-            <div id="small-dialog4" class="mfp-hide">
-                <h3>Feedback</h3>
-                <div class="feedback-grids">
-                    <div class="feedback-grid">
-                        <p>Suspendisse tristique magna ut urna pellentesque, ut egestas velit faucibus. Nullam mattis lectus ullamcorper dui dignissim, sit amet egestas orci ullamcorper.</p>
-                    </div>
-                    <div class="button-bottom">
-                        <p><a href="#small-dialog" class="play-icon popup-with-zoom-anim">Sign in</a> to get started.</p>
-                    </div>
-                </div>
-            </div>
         </div>
         <div class="signin">
             <a href="#small-dialog" class="play-icon popup-with-zoom-anim">Sign In</a>
@@ -52,9 +42,9 @@
 
                 <div class="signup">
                     <form>
-                        <input type="text" class="email" placeholder="Enter email / mobile" required="required" pattern="([\w-\.]+@([\w-]+\.)+[\w-]{2,4})?" />
+                        <input type="text" placeholder="Enter email / mobile" required="required" pattern="([\w-\.]+@([\w-]+\.)+[\w-]{2,4})?" />
                         <input type="password" placeholder="Password" required="required" pattern=".{6,}" title="Minimum 6 characters required" autocomplete="off" />
-                        <input type="submit" value="LOGIN" />
+                        <input type="submit" value="Login " />
                     </form>
                     <div class="forgot">
                         <a href="#">Forgot password ?</a>
